Restructure files, remove Card references, remove unused code
