--- ui/src/views/renders/new/Controls/cards/ControlsCardTexture/index.tsx
@@ -1,125 +1,0 @@
-import { ControlsCard } from '../ControlsCard';
-import { TextInputControl } from '@/components/form-controls/TextInputControl';
-import { TextArrayInputControl } from '@/components/form-controls/TextArrayInputControl';
-import { InfoIconAdditionalInfo } from '@/views/renders/new/Controls/icons/InfoIconAdditionalInfo';
-import { NestedTextureHeader } from './NestedTextureHeader';
-import { getTextureData } from '@/utils/render/texture';
-import { fixReferences, renameTexture } from '@/utils/render/utils';
-import type { NormalizedRenderConfig } from '@/utils/render/config';
-import type { RenderForm } from '@/hooks/useRenderForm';
-import { useSelector } from '@tanstack/react-store';
-import { Separator } from '@/components/Separator';
-
-export type ControlsCardTextureProps = {
-  form: RenderForm;
-  textureName: string;
-};
-
-export function ControlsCardTexture(props: ControlsCardTextureProps) {
-  const { form, textureName } = props;
-
-  const renderConfig = useSelector(form.store, (state) => state.values);
-
-  const { data: textureData } = getTextureData(renderConfig, textureName);
-
-  function handleRename(newName: string) {
-    const trimmed = newName.trim();
-    if (!trimmed || trimmed === textureName) {
-      return;
-    }
-
-    const renamed = renameTexture(renderConfig, textureName, trimmed);
-
-    form.setFieldValue('textures', renamed.textures);
-    form.setFieldValue('materials', renamed.materials);
-    form.setFieldValue('geometrics', renamed.geometrics);
-    form.setFieldValue('scenes', renamed.scenes);
-  }
-
-  function handleDeleteTexture(name: string) {
-    const newTextures = { ...renderConfig.textures };
-    delete newTextures[name];
-
-    const newConfig: NormalizedRenderConfig = {
-      ...renderConfig,
-      textures: newTextures,
-    };
-
-    const fixed = fixReferences(newConfig);
-    form.setFieldValue('textures', fixed.textures);
-    form.setFieldValue('materials', fixed.materials);
-    form.setFieldValue('geometrics', fixed.geometrics);
-    form.setFieldValue('scenes', fixed.scenes);
-  }
-
-  function SubTexture({ name }: { name: string }) {
-    return (
-      <>
-        <NestedTextureHeader textureName={name} renderConfig={renderConfig} />
-        {renderControls(name)}
-      </>
-    );
-  }
-
-  function renderControls(name: string) {
-    const { data } = getTextureData(renderConfig, name);
-
-    switch (data.type) {
-      case 'checker': {
-        return (
-          <>
-            <TextInputControl
-              form={form}
-              fieldName={`textures.${name}.scale`}
-              label="Scale"
-              valueLabel="scale"
-              type="number"
-            />
-            <Separator />
-            <SubTexture name={data.even_texture} />
-            <Separator />
-            <SubTexture name={data.odd_texture} />
-          </>
-        );
-      }
-      // case 'image': {
-      //   return <p className="text-sm text-zinc-500">Image texture — TODO</p>;
-      // }
-      case 'color': {
-        return (
-          <TextArrayInputControl
-            form={form}
-            fieldName={`textures.${name}.color`}
-            label="Color"
-            valueLabels={['red', 'green', 'blue']}
-            type="number"
-            unenforcedStep={0.01}
-            labelSuffix={
-              <InfoIconAdditionalInfo
-                info={[
-                  'Color values are typically between 0 and 1. For example, pure white would be [1, 1, 1].',
-                  'Values can be set outside of this range, and will be affected by the "Use Scaling Truncation" parameter.',
-                  'In the case of emissive textures, these values linearly correspond to the intensity of the emitted light.',
-                ]}
-              />
-            }
-          />
-        );
-      }
-      default:
-        return <h6 className="text-sm">Unimplemented texture: {data.type} (sorry!)</h6>;
-    }
-  }
-
-  return (
-    <ControlsCard
-      leftLabel={textureName}
-      onRename={handleRename}
-      rightLabel={textureData.type}
-      rightLabelStyle="light"
-      onDelete={() => handleDeleteTexture(textureName)}
-    >
-      <div className="flex w-full flex-col gap-2 p-4">{renderControls(textureName)}</div>
-    </ControlsCard>
-  );
-}
